Reorganize functions and callbacks
