+++ includes/metabox.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Metabox related code here
+ *
+ * @since 1.4.2
+ */
+
+namespace ucare;
+
+
+function render_ticket_category_metabox() {
+echo 'sdf';
+}
