+++ includes/post-support_ticket.php
@@ -7,6 +7,7 @@ use smartcat\form\Form;
 use smartcat\form\SelectBoxField;
 use smartcat\post\FormMetaBox;
 
+
 function register_ticket_post_type() {
 
     $labels = array(
@@ -38,25 +39,26 @@ function register_ticket_post_type() {
     );
 
     $args = array(
-        'label'               => __( 'Support Ticket', 'ucare' ),
-        'description'         => __( 'Tickets for support requests', 'ucare' ),
-        'labels'              => $labels,
-        'supports'            => array( 'editor', 'comments', 'title' ),
-        'hierarchical'        => false,
-        'public'              => false,
-        'show_ui'             => true,
-        'show_in_menu'        => false,
-        'menu_position'       => 10,
-        'menu_icon'           => 'dashicons-sos',
-        'show_in_admin_bar'   => false,
-        'show_in_nav_menus'   => false,
-        'can_export'          => true,
-        'has_archive'         => false,
-        'exclude_from_search' => true,
-        'publicly_queryable'  => false,
-        'capability_type'     => array( 'support_ticket', 'support_tickets' ),
-        'feeds'               => null,
-        'map_meta_cap'        => true
+        'label'                => __( 'Support Ticket', 'ucare' ),
+        'description'          => __( 'Tickets for support requests', 'ucare' ),
+        'labels'               => $labels,
+        'supports'             => array( 'editor', 'comments', 'title' ),
+        'hierarchical'         => false,
+        'public'               => false,
+        'show_ui'              => true,
+        'show_in_menu'         => false,
+        'menu_position'        => 10,
+        'menu_icon'            => 'dashicons-sos',
+        'show_in_admin_bar'    => false,
+        'show_in_nav_menus'    => false,
+        'can_export'           => true,
+        'has_archive'          => false,
+        'exclude_from_search'  => true,
+        'publicly_queryable'   => false,
+        'capability_type'      => array( 'support_ticket', 'support_tickets' ),
+        'feeds'                => null,
+        'map_meta_cap'         => true,
+        'register_meta_box_cb' => 'ucare\add_ticket_metaboxes'
     );
 
     register_post_type( 'support_ticket', $args );
@@ -66,38 +68,18 @@ function register_ticket_post_type() {
 add_action( 'init', 'ucare\register_ticket_post_type' );
 
 
-function register_category_taxonomy() {
 
-    $labels = array(
-        'name'                       => _x( 'Ticket Categories', 'taxonomy general name', 'ucare' ),
-        'singular_name'              => _x( 'Ticket Category', 'taxonomy singular name', 'ucare' ),
-        'all_items'                  => __( 'All Categories', 'ucare' ),
-        'edit_item'                  => __( 'Edit Category', 'ucare' ),
-        'view_item'                  => __( 'View Category', 'ucare' ),
-        'update_item'                => __( 'Update Category', 'ucare' ),
-        'add_new_item'               => __( 'Add New Category', 'ucare' ),
-        'new_item_name'              => __( 'New Category', 'ucare' ),
-        'parent_item'                => __( 'Parent Category', 'ucare' ),
-        'parent_item_colon'          => __( 'Parent Category:', 'ucare' ),
-        'search_items'               => __( 'Search Categories', 'ucare' ),
-        'popular_items'              => __( 'Popular Categories', 'ucare' ),
-        'not_found'                  => __( 'No categories found', 'ucare' ),
-        'add_or_remove_items'        => __( 'Add or remove categories', 'ucare' ),
-        'choose_from_most_used'      => __( 'Choose from the most used categories', 'ucare' ),
-        'separate_items_with_commas' => __( 'Separate categories with commas', 'ucare' )
-    );
+function add_ticket_metaboxes() {
 
-    $args = array(
-        'label'        => __( 'Categories', 'ucare' ),
-        'labels'       => $labels,
-        'hierarchical' => false
-    );
+    // Remove the original category metabox
+    remove_meta_box( 'tagsdiv-ticket_category', 'support_ticket', 'side' );
 
-    register_taxonomy( 'ticket_category', 'support_ticket', $args );
+    // Add category metabox
+    add_meta_box( 'ticket-category', __( 'Category', 'ucare' ), 'ucare\render_ticket_category_metabox', 'support_ticket', 'side' );
 
 }
 
-add_action( 'init', 'ucare\register_category_taxonomy' );
+
 
 
 function force_menu_expand( $file ) {
